+++ src/modules/chanmodes/limit.c
@@ -48,7 +48,7 @@ int cmode_limit_is_ok(Client *client, Channel *channel, char mode, const char *p
 void *cmode_limit_put_param(void *r_in, const char *param);
 const char *cmode_limit_get_param(void *r_in);
 const char *cmode_limit_conv_param(const char *param_in, Client *client, Channel *channel);
-int cmode_limit_free_param(void *r);
+int cmode_limit_free_param(void *r, int soft);
 void *cmode_limit_dup_struct(void *r_in);
 int cmode_limit_sjoin_check(Channel *channel, void *ourx, void *theirx);
 int transform_channel_limit(const char *param);
@@ -159,7 +159,7 @@ const char *cmode_limit_conv_param(const char *param, Client *client, Channel *c
 	return retbuf;
 }
 
-int cmode_limit_free_param(void *r)
+int cmode_limit_free_param(void *r, int soft)
 {
 	safe_free(r);
 	return 0;